+++ server/src/main/java/org/allaymc/server/block/type/BlockTypeInitializer.java
@@ -504,13 +504,18 @@ public final class BlockTypeInitializer {
         BlockTypes.WAXED_EXPOSED_CUT_COPPER_STAIRS = buildCopperStairs(BlockId.WAXED_EXPOSED_CUT_COPPER_STAIRS, OxidationLevel.EXPOSED, cutCopperStairs);
         BlockTypes.WAXED_WEATHERED_CUT_COPPER_STAIRS = buildCopperStairs(BlockId.WAXED_WEATHERED_CUT_COPPER_STAIRS, OxidationLevel.WEATHERED, cutCopperStairs);
         BlockTypes.WAXED_OXIDIZED_CUT_COPPER_STAIRS = buildCopperStairs(BlockId.WAXED_OXIDIZED_CUT_COPPER_STAIRS, OxidationLevel.OXIDIZED, cutCopperStairs);
+
+        BlockTypes.RESIN_BRICK_STAIRS = buildStairs(BlockId.RESIN_BRICK_STAIRS);
     }
 
     private static BlockType<BlockStairsBehavior> buildStairs(BlockId id) {
-        return stairsBuilder(BlockStairsBehaviorImpl.class, id).setBaseComponentSupplier(BlockStairsBaseComponentImpl::new).build();
+        return stairsBuilder(BlockStairsBehaviorImpl.class, id).build();
     }
 
-    private static BlockType<BlockCopperStairsBehavior> buildCopperStairs(BlockId id, OxidationLevel oxidationLevel, BiFunction<OxidationLevel, Boolean, BlockType<? extends BlockOxidationComponent>> blockTypeFunction) {
+    private static BlockType<BlockCopperStairsBehavior> buildCopperStairs(
+            BlockId id, OxidationLevel oxidationLevel,
+            BiFunction<OxidationLevel, Boolean, BlockType<? extends BlockOxidationComponent>> blockTypeFunction
+    ) {
         return stairsBuilder(BlockCopperStairsBehaviorImpl.class, id)
                 .addComponent(new BlockOxidationComponentImpl(oxidationLevel, blockTypeFunction))
                 .build();
@@ -520,7 +525,8 @@ public final class BlockTypeInitializer {
         return AllayBlockType
                 .builder(clazz)
                 .vanillaBlock(id)
-                .setProperties(BlockPropertyTypes.UPSIDE_DOWN_BIT, BlockPropertyTypes.WEIRDO_DIRECTION);
+                .setProperties(BlockPropertyTypes.UPSIDE_DOWN_BIT, BlockPropertyTypes.WEIRDO_DIRECTION)
+                .setBaseComponentSupplier(BlockStairsBaseComponentImpl::new);
     }
 
     public static void initSlab() {
@@ -1209,14 +1215,6 @@ public final class BlockTypeInitializer {
                 .build();
     }
 
-    public static BlockType<BlockTrapdoorBehavior> buildTrapdoor(BlockId blockId) {
-        return AllayBlockType.builder(BlockTrapdoorBehaviorImpl.class)
-                .vanillaBlock(blockId)
-                .setProperties(BlockPropertyTypes.DIRECTION_4, BlockPropertyTypes.OPEN_BIT, BlockPropertyTypes.UPSIDE_DOWN_BIT)
-                .setBaseComponentSupplier(BlockTrapdoorBaseComponentImpl::new)
-                .build();
-    }
-
     public static void initTrapdoor() {
         BlockTypes.ACACIA_TRAPDOOR = buildTrapdoor(BlockId.ACACIA_TRAPDOOR);
         BlockTypes.BAMBOO_TRAPDOOR = buildTrapdoor(BlockId.BAMBOO_TRAPDOOR);
@@ -1236,6 +1234,41 @@ public final class BlockTypeInitializer {
                 .setProperties(BlockPropertyTypes.DIRECTION_4, BlockPropertyTypes.OPEN_BIT, BlockPropertyTypes.UPSIDE_DOWN_BIT)
                 .setBaseComponentSupplier(BlockIronTrapdoorBaseComponentImpl::new)
                 .build();
+
+        BiFunction<OxidationLevel, Boolean, BlockType<? extends BlockOxidationComponent>> copperTrapdoor = (level, waxed) -> switch (level) {
+            case UNAFFECTED -> waxed ? BlockTypes.WAXED_COPPER_TRAPDOOR : BlockTypes.COPPER_TRAPDOOR;
+            case EXPOSED -> waxed ? BlockTypes.WAXED_EXPOSED_COPPER_TRAPDOOR : BlockTypes.EXPOSED_COPPER_TRAPDOOR;
+            case WEATHERED -> waxed ? BlockTypes.WAXED_WEATHERED_COPPER_TRAPDOOR : BlockTypes.WEATHERED_COPPER_TRAPDOOR;
+            case OXIDIZED -> waxed ? BlockTypes.WAXED_OXIDIZED_COPPER_TRAPDOOR : BlockTypes.OXIDIZED_COPPER_TRAPDOOR;
+        };
+        BlockTypes.COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.COPPER_TRAPDOOR, OxidationLevel.UNAFFECTED, copperTrapdoor);
+        BlockTypes.EXPOSED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.EXPOSED_COPPER_TRAPDOOR, OxidationLevel.EXPOSED, copperTrapdoor);
+        BlockTypes.WEATHERED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.WEATHERED_COPPER_TRAPDOOR, OxidationLevel.WEATHERED, copperTrapdoor);
+        BlockTypes.OXIDIZED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.OXIDIZED_COPPER_TRAPDOOR, OxidationLevel.OXIDIZED, copperTrapdoor);
+        BlockTypes.WAXED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.WAXED_COPPER_TRAPDOOR, OxidationLevel.UNAFFECTED, copperTrapdoor);
+        BlockTypes.WAXED_EXPOSED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.WAXED_EXPOSED_COPPER_TRAPDOOR, OxidationLevel.EXPOSED, copperTrapdoor);
+        BlockTypes.WAXED_WEATHERED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.WAXED_WEATHERED_COPPER_TRAPDOOR, OxidationLevel.WEATHERED, copperTrapdoor);
+        BlockTypes.WAXED_OXIDIZED_COPPER_TRAPDOOR = buildCopperTrapdoor(BlockId.WAXED_OXIDIZED_COPPER_TRAPDOOR, OxidationLevel.OXIDIZED, copperTrapdoor);
+    }
+
+    public static BlockType<BlockTrapdoorBehavior> buildTrapdoor(BlockId blockId) {
+        return trapdoorBuilder(BlockTrapdoorBehaviorImpl.class, blockId).build();
+    }
+
+    public static BlockType<BlockCopperTrapdoorBehavior> buildCopperTrapdoor(
+            BlockId id, OxidationLevel oxidationLevel,
+            BiFunction<OxidationLevel, Boolean, BlockType<? extends BlockOxidationComponent>> blockTypeFunction
+    ) {
+        return trapdoorBuilder(BlockCopperStairsBehaviorImpl.class, id)
+                .addComponent(new BlockOxidationComponentImpl(oxidationLevel, blockTypeFunction))
+                .build();
+    }
+
+    public static <T extends BlockBehavior> AllayBlockType.Builder trapdoorBuilder(Class<T> clazz, BlockId id) {
+        return AllayBlockType.builder(clazz)
+                .vanillaBlock(id)
+                .setProperties(BlockPropertyTypes.DIRECTION_4, BlockPropertyTypes.OPEN_BIT, BlockPropertyTypes.UPSIDE_DOWN_BIT)
+                .setBaseComponentSupplier(BlockTrapdoorBaseComponentImpl::new);
     }
 
     public static void initSponge() {
@@ -1327,7 +1360,6 @@ public final class BlockTypeInitializer {
                 .builder(BlockBrewingStandBehaviorImpl.class)
                 .vanillaBlock(BlockId.BREWING_STAND)
                 .bindBlockEntity(BlockEntityTypes.BREWING_STAND)
-                .setBaseComponentSupplier(BlockBrewingStandBaseComponentImpl::new)
                 .setProperties(BlockPropertyTypes.BREWING_STAND_SLOT_A_BIT, BlockPropertyTypes.BREWING_STAND_SLOT_B_BIT, BlockPropertyTypes.BREWING_STAND_SLOT_C_BIT)
                 .build();
     }
@@ -2002,6 +2034,33 @@ public final class BlockTypeInitializer {
                 .vanillaBlock(blockId)
                 .setProperties(BlockPropertyTypes.GROWTH, BlockPropertyTypes.MINECRAFT_CARDINAL_DIRECTION)
                 .setBaseComponentSupplier(BlockPlantPileBaseComponentImpl::new)
+                .build();
+    }
+
+    public static void initCopperBars() {
+        BiFunction<OxidationLevel, Boolean, BlockType<? extends BlockOxidationComponent>> copperBars = (level, waxed) -> switch (level) {
+            case UNAFFECTED -> waxed ? BlockTypes.WAXED_COPPER_BARS : BlockTypes.COPPER_BARS;
+            case EXPOSED -> waxed ? BlockTypes.WAXED_EXPOSED_COPPER_BARS : BlockTypes.EXPOSED_COPPER_BARS;
+            case WEATHERED -> waxed ? BlockTypes.WAXED_WEATHERED_COPPER_BARS : BlockTypes.WEATHERED_COPPER_BARS;
+            case OXIDIZED -> waxed ? BlockTypes.WAXED_OXIDIZED_COPPER_BARS : BlockTypes.OXIDIZED_COPPER_BARS;
+        };
+        BlockTypes.COPPER_BARS = buildCopperBars(BlockId.COPPER_BARS, OxidationLevel.UNAFFECTED, copperBars);
+        BlockTypes.EXPOSED_COPPER_BARS = buildCopperBars(BlockId.EXPOSED_COPPER_BARS, OxidationLevel.EXPOSED, copperBars);
+        BlockTypes.WEATHERED_COPPER_BARS = buildCopperBars(BlockId.WEATHERED_COPPER_BARS, OxidationLevel.WEATHERED, copperBars);
+        BlockTypes.OXIDIZED_COPPER_BARS = buildCopperBars(BlockId.OXIDIZED_COPPER_BARS, OxidationLevel.OXIDIZED, copperBars);
+        BlockTypes.WAXED_COPPER_BARS = buildCopperBars(BlockId.WAXED_COPPER_BARS, OxidationLevel.UNAFFECTED, copperBars);
+        BlockTypes.WAXED_EXPOSED_COPPER_BARS = buildCopperBars(BlockId.WAXED_EXPOSED_COPPER_BARS, OxidationLevel.EXPOSED, copperBars);
+        BlockTypes.WAXED_WEATHERED_COPPER_BARS = buildCopperBars(BlockId.WAXED_WEATHERED_COPPER_BARS, OxidationLevel.WEATHERED, copperBars);
+        BlockTypes.WAXED_OXIDIZED_COPPER_BARS = buildCopperBars(BlockId.WAXED_OXIDIZED_COPPER_BARS, OxidationLevel.OXIDIZED, copperBars);
+    }
+
+    public static BlockType<BlockCopperBarsBehavior> buildCopperBars(
+            BlockId id, OxidationLevel oxidationLevel,
+            BiFunction<OxidationLevel, Boolean, BlockType<? extends BlockOxidationComponent>> blockTypeFunction
+    ) {
+        return AllayBlockType.builder(BlockCopperBarsBehaviorImpl.class)
+                .vanillaBlock(id)
+                .addComponent(new BlockOxidationComponentImpl(oxidationLevel, blockTypeFunction))
                 .build();
     }
 }